project/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-crispy-forms==1.12.0
docopt==0.6.2
idna==2.10
Js2Py==0.71
packaging==20.9
pipwin==0.5.1
psycopg2-binary==2.9.1
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
